backend/requirements.txt
alembic
Pillow
psycopg2-binary
requests
SQLAlchemy
tornado
tornado-sqlalchemy
